[tests]: add asserts for api dto tests
